% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_mode.R
\name{context.mode}
\alias{context.mode}
\title{Building context mode}
\usage{
context.mode(data, qt.gene = 0.4, qt.cell = 0.5, nThreads = 1)
}
\arguments{
\item{data}{A expression count matrix. The rows correspond to genes and
the columns correspond to cells.}

\item{qt.gene}{A numeric value between 0 and 1 (default: 0.4) indicating the
top percent of expressed genes to be reserved for buliding a context-specific
gene-gene network. Used only if \code{network} = "context".}

\item{qt.cell}{A numeric value between 0 and 1 (default: 0.5) indicating the
top percent of expressed cells to be reserved for buliding a context-specific
gene-gene network. Used only if \code{network} = "context".}

\item{nThreads}{The number of cores to use. Default is 1.}
}
\value{
A context-specific gene-gene network.
}
\description{
a context-specific gene-gene network is constructed from
                 the scRNA-seq data by WGCNA package.
}
\examples{
#Loading example scRNA-seq data.
load(system.file("data","yan.Rdata",package = "scNPF"))
exp.data <- yan$data
dim(exp.data)
label <- yan$label

#Parallel
context.network <- context.mode(data=exp.data,nThreads=8)
dim(context.network)
class(context.network)
}
