% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap}
\usage{
plotHeatmap(dists, label)
}
\arguments{
\item{dists}{A distance or similarity matrix.}

\item{label}{True labels of sample.}
}
\value{
This function returns a heatmap object invisibly.
}
\description{
This function plots a heatmap figure form a distance or
  similarity matrix. The colors denote the sample types based on \code{label}.
}
\examples{
library(gplots)
library(RColorBrewer)
load(system.file("data","yan.Rdata",package = "scNPF"))
data.dist <- as.matrix(dist(t(yan$data)))
#dev.off()
plotHeatmap(data.dist,yan$label)
}
